+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.1.35 - 2026-06-08
 ### Fixed
 - Retry mechanism for failing Nitecrawler serial comm
+- Share invalid chars across platforms
 
 ## 1.1.34 - 2026-06-02
 ### Added